# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Test Go Go Ahead",
        "title": "Test Go Go Ahead",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/currency": {
            "get": {
                "description": "Get currency by params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Get currency",
                "operationId": "get-currency-by-params",
                "parameters": [
                    {
                        "type": "string",
                        "example": "22.05.2024",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "USD",
                        "name": "val",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Currency"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Currency": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2024-05-22T00:00:00Z"
                },
                "nominal": {
                    "type": "integer",
                    "example": 1
                },
                "val": {
                    "type": "string",
                    "example": "USD"
                },
                "value": {
                    "type": "number",
                    "example": 90.22
                }
            }
        },
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}